멤버별 관심 분야 db저장

+++ uhang/src/main/java/uhang/uhang/login/domain/repository/MemberRepository.java
@@ -16,6 +16,6 @@ public interface MemberRepository extends JpaRepository<Member, Integer> {
     public Member findByMemberEmail(String memberEmail);
     public boolean existsByMemberEmail(String email);
 
-    public Member findByMemberId(int memberId);
+    public Member findByMemberId(Long memberId);
 }
